requirements-all.txt
alembic==1.9.0
aniso8601==9.0.1
annotated-types==0.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
coverage==6.3.3
cryptography==38.0.4
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
Flask==2.2.2
Flask-RESTful==0.3.9
flask-restful-swagger==0.20.2
fonttools==4.38.0
future==0.18.3
greenlet==2.0.1
h11==0.13.0
html2text==2020.1.16
idna==3.3
ImageHash==4.3.1
importlib-metadata==5.1.0
importlib-resources==5.10.1
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
kiwisolver==1.4.4
lxml==4.6.5
Mako==1.2.4
Markdown==3.4.1
markdown2==2.4.6
MarkupSafe==2.1.1
matplotlib==3.6.2
mdx-truly-sane-lists==1.3
mpmath==1.2.1
mysql-connector-python==8.0.29
networkx==2.8.8
numpy==1.24.1
oauthlib==3.2.2
oic==1.6.1
outcome==1.1.0
packaging==21.3
pandas==1.5.2
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-settings==2.0.3
pydot==1.4.2
pygraphviz==1.10
pyjwkest==1.4.2
PyJWT==2.6.0
pyop==3.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.1
pytest-cov==3.0.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python3-saml==1.14.0
pytz==2022.6
PyVirtualDisplay==3.0
PyWavelets==1.4.1
requests==2.28.1
scipy==1.10.0
selenium==4.6.1
semantic-version==2.10.0
sentry-sdk==1.14.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sympy==1.11.1
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.7.1
urllib3==1.26.14
waitress==2.1.2
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.1.0
xmlsec==1.3.13
zipp==3.11.0
